# documentation/swagger.yaml
swagger: "2.0"
info: 
  version: "1.0.0"
  title: GameLabz
  description: |
    ## Required software:

    - Java 1.8+
    - PostgreSQL 9+
    - Apache Maven 3+
    
    ## Build:
    CD to the pom folder
    
    ```
    mvn clean package -DskipTests
    ```
    The jar will be in the /target folder.
    
    This SQL script has to be executed once when the server is deployed for the first time:
    
    ```
    drop database if exists "gameLabz";
    create database "gameLabz";
    ```
    
    The required tables will be created by the application.
    
    ## Usage: 
    ```
    java -jar gameLabz-1.0-SNAPSHOT.jar [parameters]
    Parameters:
    --server.port: the server port, default: 8080
    --database.username: the username for the PostgreSQL database, mandatory
    --database.password: the password for the PostgreSQL database, mandatory
    ```
  contact:
    name: Razman Tamas
host: localhost:8080
basePath: /gameLabz
schemes:
  - http
paths:
...
  /admins/{userName}/games:
    post:
      tags:
        - games
      summary: An admin creates a game
      operationId: addGame
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: userName
          description: Username of the calling admin user who wants to create a game
          required: true
          type: string
        - in: header
          name: session_id
          description: It is used to identify the user's session
          required: true
          type: string
        - in: body
          name: gameBody
          description: New game that needs to be added
          required: true
          schema:
            $ref: "#/definitions/Game"
      responses:
        201:
          description: A new game has been successfully created
          schema:
            $ref: "#/definitions/GameWithId"
        400:
          description: Invalid input
          headers:
            error_message:
              type: string
              description: The error message
        401:
          description: Admin's session is invalid
          headers:
            error_message:
              type: string
              description: The error message
        403:
          description: Only admins can create games
          headers:
            error_message:
              type: string
              description: The error message
          
    get:
      tags:
        - games
      summary: An admin lists games by certain filters
      operationId: getGames
      produces:
        - application/json
      parameters:
        - in: path
          name: userName
          description: Username of the calling admin user who wants to list games
          required: true
          type: string
        - in: header
          name: session_id
          description: It is used to identify the user's session
          required: true
          type: string
        - in: query
          name: gameName
          description: To filter games by their name
          required: false
          type: string
        - in: query
          name: hasOffer
          description: To filter games which are/aren't in offers 
          required: false
          type: boolean
        - in: query
          name: genre
          description: To filter games by their genre
          required: false
          type: string
        - in: query
          name: releasedBefore
          description: To filter games which were released before the given date
          required: false
          type: integer
          format: int64
        - in: query
          name: releasedAfter
          description: To filter games which were released after the given date
          required: false
          type: integer
          format: int64
      responses:
        200:
          description: Listing games was successful
          schema:
            type: array
            items: 
              $ref: "#/definitions/GameWithId"
        400:
          description: Invalid parameters
          headers:
            error_message:
              type: string
              description: The error message
        401:
          description: Admin's session is invalid
          headers:
            error_message:
              type: string
              description: The error message
        403:
          description: Only admins can list games
          headers:
            error_message:
              type: string
              description: The error message
  /admins/{userName}/games/{gameId}:
    put:
      tags:
        - games
      summary: An admin updates an existing game
      description: "Finds a game by id and updates it. The following fields can be updated: name, description, release date and genre"
      operationId: updateGame
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: userName
          description: Username of the calling admin user who wants to update the game
          required: true
          type: string
        - in: path
          name: gameId
          description: Id of the game which needs to be updated
          required: true
          type: integer
          format: int64
        - in: header
          name: session_id
          description: It is used to identify the user's session
          required: true
          type: string
        - in: body
          name: gameBody
          description: Game with new fields to update the old one
          required: true
          schema:
            $ref: "#/definitions/Game"
      responses:
        200:
          description: Updating game was successful
          schema:
            $ref: "#/definitions/GameWithId"
        400:
          description: Invalid parameters
          headers:
            error_message:
              type: string
              description: The error message
        401:
          description: Admin's session is invalid
          headers:
            error_message:
              type: string
              description: The error message
        403:
          description: Only admins can update games
          headers:
            error_message:
              type: string
              description: The error message
          
      
      
#----------------------- GAME OFFER ------------------------
...
definitions:
...
  Game:
      type: object
      properties:
        name:
          type: string
        description: 
          type: string
        releaseDate:
          type: integer
          format: int64
        gameGenre:
          type: string
          enum:
            - RPG
            - FPS
            - MOBA
  GameWithId:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description: 
          type: string
        releaseDate:
          type: integer
          format: int64
        gameGenre:
          type: string
          enum:
            - RPG
            - FPS
            - MOBA
...